requirements.txt
torch>=2.2.0
torchvision>=0.17.0
opencv-python
numpy
pandas
scikit-learn
matplotlib
seaborn
pillow
tqdm
albumentations
pyyaml
jupyter
scikit-image
tensorboard
